price calculation and purchase flows

+++ src/main/java/com/zakharovmm/noveotestassignment/entity/Product.java
@@ -2,18 +2,23 @@ package com.zakharovmm.noveotestassignment.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
-import static jakarta.persistence.GenerationType.IDENTITY;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import java.math.BigDecimal;
 import lombok.Data;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
 
 @Data
 @Entity
 @Table(name = "products")
 public class Product {
+
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
+
     private String name;
-    private float price;
+
+    private BigDecimal price;
 }
